refactor: vendor dependencies and migrate to numpy 2.0 (#15)
* Squashed 'vendor/simple-lama-inpainting/' content from commit 0790c45

git-subtree-dir: vendor/simple-lama-inpainting
git-subtree-split: 0790c455160991a6a1417723a68d3d48e3c859e5

* Squashed 'vendor/cr-renderer/' content from commit a17e1fb

git-subtree-dir: vendor/cr-renderer
git-subtree-split: a17e1fb9a732959dd6c21ae168d5ee09d74bd64f

* refactor: vendor dependencies and migrate to numpy 2.0

This commit refactors LayerD to vendor unmaintained dependencies using
UV workspace and git subtrees, while migrating to numpy 2.0 for better
compatibility and performance.

## Changes

### Git Subtrees
- Added `vendor/simple-lama-inpainting/` from upstream main branch
- Added `vendor/cr-renderer/` at revision a17e1fb
- Both packages vendored using git subtree for maintainability

### UV Workspace Configuration
- Configured UV workspace with 3 members (root + 2 vendored packages)
- Updated `[tool.uv.sources]` to use workspace references
- Migrated to `[dependency-groups]` from `[project.optional-dependencies]`

### Numpy 2.0 Migration
- Updated numpy requirement: `numpy>=1.26` → `numpy>=2.0`
- Fixed deprecated `np.bool_` → `bool` in base.py
- Fixed numpy scalar type hint in metrics.py (wrapped np.mean in float())
- Updated both vendored packages to require numpy>=2.0
- No deprecated numpy types found in vendored code

### Vendored Package Updates
- Converted simple-lama-inpainting from Poetry to hatchling build system
- Updated cr-renderer to use `[dependency-groups]` instead of `[tool.uv.dev-dependencies]`
- Both packages now compatible with numpy 2.0

### Documentation
- Added "Vendored Dependencies" section to CLAUDE.md
- Updated README.md third-party libraries section
- Documented git subtree update workflow

## Testing
- ✅ All 4 tests pass
- ✅ Type checking passes (mypy on 28 source files)
- ✅ UV workspace resolves dependencies correctly with numpy 2.4.0



---------

+++ src/layerd/evaluation/metrics.py
@@ -12,7 +12,7 @@ class MAE:
     def __call__(self, x: np.ndarray, y: np.ndarray) -> float:
         x = x / self.normalize_value
         y = y / self.normalize_value
-        return np.mean(np.abs(x - y))
+        return float(np.mean(np.abs(x - y)))
 
 
 class SSIM:
